Move `AddExtension` into `extension` module
