--- axum/src/add_extension.rs
@@ -1,37 +1,0 @@
-// this is vendored from tower-http to reduce public dependencies
-
-use http::Request;
-use std::task::{Context, Poll};
-use tower_service::Service;
-
-/// Middleware for adding some shareable value to [request extensions].
-///
-/// See [Sharing state with handlers](index.html#sharing-state-with-handlers)
-/// for more details.
-///
-/// [request extensions]: https://docs.rs/http/latest/http/struct.Extensions.html
-#[derive(Clone, Copy, Debug)]
-pub struct AddExtension<S, T> {
-    pub(crate) inner: S,
-    pub(crate) value: T,
-}
-
-impl<ResBody, S, T> Service<Request<ResBody>> for AddExtension<S, T>
-where
-    S: Service<Request<ResBody>>,
-    T: Clone + Send + Sync + 'static,
-{
-    type Response = S::Response;
-    type Error = S::Error;
-    type Future = S::Future;
-
-    #[inline]
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        self.inner.poll_ready(cx)
-    }
-
-    fn call(&mut self, mut req: Request<ResBody>) -> Self::Future {
-        req.extensions_mut().insert(self.value.clone());
-        self.inner.call(req)
-    }
-}
