+++ axum/src/middleware/mod.rs
@@ -5,7 +5,7 @@
 mod from_fn;
 
 pub use self::from_fn::{from_fn, FromFn, FromFnLayer, Next};
-pub use crate::add_extension::AddExtension;
+pub use crate::extension::AddExtension;
 
 pub mod future {
     //! Future types.
